requirement.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
idna==2.8
python-firebase==1.2.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
